# api/v1/swagger.yaml
...
swagger: "2.0"
...
info:
  version: 1.0.0
  title: Refocus API
  # description: >-
  #   A short description of the application. GFM syntax can be used for rich text
  #   representation.
  # termsOfService: The Terms of Service for the API.
  # contact:
  #   name: The identifying name of the contact person/organization.
  #   url: http://focus.salesforce.com/contact
  #   email: focusapi@salesforce.com
  # license:
  #   name: The license name used for the API.
  #   url: http://focus.salesforce.com/license

# =============================================================================
host: localhost:10010
...
basePath: /v1
...
schemes:
  - https
  - http

# =============================================================================
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
# =============================================================================
consumes:
  - application/json

# =============================================================================
produces:
  - application/json

# =============================================================================
paths:
...
  /ssoconfig:
    x-swagger-router-controller: ssoconfig
    get:
      security:
        - jwt: []
      summary: Retrieve SSO Config
      tags: [ ssoconfig ]
      description: >-
        Retrieves SSO Config.
      operationId: getSSOConfig
      responses:
        200:
          description: >-
            Success, returns SSO config
          schema:
            $ref: "#/definitions/SSOConfigResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    post:
      security:
        - jwt: []
      summary: Create SSO config
      tags: [ ssoconfig ]
      description: >-
        Create SSO config. POST will throw SSOConfigCreateConstraintError if
        SSOConfig already exists.
      operationId: postSSOConfig
      parameters:
        -
          name: queryBody
          in: body
          schema:
            type: object
            description: >
              Configuration parameters to enable SSO in Refocus.
            properties:
              samlEntryPoint:
                type: string
                maxLength: 2082
                description: >
                  Identity Provider endpoint.
              samlIssuer:
                type: string
                maxLength: 2082
                description: >
                  Issuer string to supply to Identity Provider.
            required:
              - samlEntryPoint
              - samlIssuer
      responses:
        201:
          description: >-
            Returns created SSO Config
          schema:
            $ref: "#/definitions/SSOConfigResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        default:
          $ref: "#/responses/genericError"
    delete:
      security:
        - jwt: []
      summary: Delete SSO Config
      tags: [ ssoconfig ]
      description: >-
        Delete SSO Config.
      operationId: deleteSSOConfig
      responses:
        200:
          description: >-
            Returns deleted SSO config
          schema:
            $ref: "#/definitions/SSOConfigResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    patch:
      security:
        - jwt: []
      summary: Update SSO Config
      tags: [ ssoconfig ]
      description: >-
        Update the specified SSO Config. If a field is not included in the query
        body, that field will not be updated.
      operationId: patchSSOConfig
      parameters:
        -
          name: queryBody
          in: body
          required: true
          schema:
            type: object
            description: >
              Configuration parameters to enable SSO in Refocus.
            properties:
              samlEntryPoint:
                type: string
                maxLength: 2082
                description: >
                  Identity Provider endpoint.
              samlIssuer:
                type: string
                maxLength: 2082
                description: >
                  Issuer string to supply to Identity Provider.
            required:
              - samlEntryPoint
              - samlIssuer
      responses:
        200:
          description: >-
            Success, returns updated SSO Config.
          schema:
            $ref: "#/definitions/SSOConfigResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    put:
      security:
        - jwt: []
      summary: Update SSO Config
      tags: [ ssoconfig ]
      description: >-
        Update the specified SSO Config. If a field is not included in the query
        body, that field will be set to null or its default value.
      operationId: putSSOConfig
      parameters:
        -
          name: queryBody
          in: body
          required: true
          schema:
            type: object
            description: >
              Configuration parameters to enable SSO in Refocus.
            properties:
              samlEntryPoint:
                type: string
                maxLength: 2082
                description: >
                  Identity Provider endpoint.
              samlIssuer:
                type: string
                maxLength: 2082
                description: >
                  Issuer string to supply to Identity Provider.
            required:
              - samlEntryPoint
              - samlIssuer
      responses:
        200:
          description: >-
            Success, returns updated SSO Config.
          schema:
            $ref: "#/definitions/SSOConfigResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

# =============================================================================
...
definitions:
...
  BadRequest:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  ErrorResponse:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  #  required:
  #    - message
  HATEOAS:
    type: object
    description: >
      Hypertext As The Engine Of Application State.
    properties:
      href:
        type: string
        maxLength: 2082
        readOnly: true
        description: >
          URL of the related HATEOAS link you can use for subsequent calls.
      rel:
        type: string
        maxLength: 255
        readOnly: true
        description: >
          Link relation that describes how this link relates to the previous call.
      method:
        type: string
        readOnly: true
        maxLength: 255
        description: >
          The HTTP method required for the related call.
    required:
      - href
      - rel
      - method
...
  ResourceNotFoundError:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors

    # properties:
    #   message:
    #     type: string
    #   resourceType:
    #     type: string
    #     enum:
    #       - Aspect
    #       - Lens
    #       - Perspective
    #       - Profile
    #       - Sample
    #       - Subject
    #       - User
    #   resourceKey:
    #     type: string
    # required:
    #   - message
...
  SSOConfigResponse:
    type: object
    description: >
      SSO Configuration parameters used to enable SSO in Refocus.
    properties:
      id:
        type: string
        readOnly: true
      samlEntryPoint:
        type: string
        readOnly: true
        maxLength: 2082
        description: >
          Required SAML Identity Provider entry point.
      samlIssuer:
        type: string
        readOnly: true
        maxLength: 2082
        description: >
          Required issuer string to supply to identity provider.
      createdBy:
        type: string
        readOnly: true
        description: >
          TODO
      createdAt:
        type: string
        readOnly: true
        description: >
          TODO
      updatedAt:
        type: string
        readOnly: true
        description: >
          TODO
      # TODO profile!
      apiLinks:
        readOnly: true
        type: array
        items:
          $ref: "#/definitions/HATEOAS"
        description: >
          Hypertext As The Engine Of Application State.
...
responses:
  400:
    description: Bad Request
    schema:
      $ref: '#/definitions/BadRequest'
...
  403:
    description: Forbidden
    schema:
      $ref: '#/definitions/ErrorResponse'
  404:
    description: Resource Not Found
    schema:
      $ref: '#/definitions/ResourceNotFoundError'
  genericError:
    description: Error
    schema:
      $ref: '#/definitions/ErrorResponse'

# =============================================================================